+++ packages/shared/src/types/index.ts
@@ -140,16 +140,25 @@ export const ArticleWidths = ['narrow', 'wide', 'unlimited'] as const;
 export type ArticleWidth = typeof ArticleWidths[number];
 export const defaultArticleWidth: ArticleWidth = 'narrow';
 
-export const ArticleFontFamilies = ['nunito', 'openDyslexic'] as const;
+export const ArticleFontFamilies = [
+  'nunito',
+  'ubuntu',
+  'lato',
+  'openDyslexic',
+] as const;
 export type ArticleFontFamily = typeof ArticleFontFamilies[number];
 export const defaultArticleFontFamily: ArticleFontFamily = 'nunito';
 export const ArticleFontFamiliesNames: Record<ArticleFontFamily, string> = {
   nunito: 'Nunito',
+  ubuntu: 'Ubuntu',
+  lato: 'Lato',
   openDyslexic: 'OpenDyslexic',
 };
 
 export const ArticleMonoFontFamilies = [
   'sourceCodePro',
+  'ubuntuMono',
+  'firaMono',
   'openDyslexicMono',
 ] as const;
 export type ArticleMonoFontFamily = typeof ArticleMonoFontFamilies[number];
@@ -161,6 +170,8 @@ export const ArticleMonoFontFamiliesNames: Record<
 > = {
   sourceCodePro: 'Source Code Pro',
   openDyslexicMono: 'OpenDyslexic Mono',
+  ubuntuMono: 'Ubuntu Mono',
+  firaMono: 'Fira Mono',
 };
 
 export const ArticleFontSizes = ['sm', 'md', 'lg', 'xl'] as const;